+++ src/apis/common/mutations.ts
@@ -1,10 +1,6 @@
-import { postFCMToken, postFiles } from './apis';
+import { postFiles } from './apis';
 import { useMutation } from '@tanstack/react-query';
 
 export const usePostFiles = () => {
   return useMutation(postFiles);
-};
-
-export const usePostFCMToken = () => {
-  return useMutation(postFCMToken);
 };
