--- src/apis/common/queries.ts
@@ -1,6 +1,0 @@
-import { Keys, getNotification } from '.';
-import { useSuspenseQuery } from '@suspensive/react-query';
-
-export const useGetNotifications = () => {
-  return useSuspenseQuery(Keys.getNotifications(), getNotification);
-};
