+++ src/apis/notifications/keys.ts
@@ -1,0 +1,3 @@
+export const Keys = {
+  getNotifications: () => ['notifications'],
+} as const;
