+++ src/apis/notifications/mutations.ts
@@ -1,0 +1,6 @@
+import { postFCMToken } from './apis';
+import { useMutation } from '@tanstack/react-query';
+
+export const usePostFCMToken = () => {
+  return useMutation(postFCMToken);
+};
