+++ src/apis/notifications/queries.ts
@@ -1,0 +1,6 @@
+import { Keys, getNotification } from '.';
+import { useSuspenseQuery } from '@suspensive/react-query';
+
+export const useGetNotifications = () => {
+  return useSuspenseQuery(Keys.getNotifications(), getNotification);
+};
