+++ src/apis/notifications/type.ts
@@ -1,0 +1,1 @@
+export type Type = {};
