+++ src/app/[lng]/(sub)/notification/components/NotificationSection.tsx
@@ -1,10 +1,11 @@
 'use client';
 import NotiCard from './NotiCard.client';
-import { useGetNotifications } from '@/apis/common';
+import { getNotification, useGetNotifications } from '@/apis/notifications';
 import { useTranslation } from '@/app/i18n/client';
 import { Icon } from '@/components/Icon';
 import { Flex } from '@/components/Layout';
 import { ItemList } from '@/components/List';
+import { useDidMount } from '@/hooks/common/useDidMount';
 import { ComponentProps } from 'react';
 
 const DUMMY_DATA: Array<ComponentProps<typeof NotiCard>['notiData']> = [
@@ -36,7 +37,7 @@ const DUMMY_DATA: Array<ComponentProps<typeof NotiCard>['notiData']> = [
 
 export default function NotificationSection() {
   const { i18n } = useTranslation('common');
-  // const { data } = useGetNotifications();
+  const { data } = useGetNotifications();
 
   return (
     <ItemList
