+++ src/app/[lng]/(sub)/notification/page.tsx
@@ -1,6 +1,6 @@
 import NotificationHeader from './components/NotificationHeader';
 import NotificationSection from './components/NotificationSection';
-import { Keys, getNotification } from '@/apis/common';
+import { Keys, getNotification } from '@/apis/notifications';
 import { RejectedFallback } from '@/components/ErrorBoundary';
 import { Loading } from '@/components/Loading';
 import { PageAnimation } from '@/components/PageAnimation';
@@ -13,7 +13,7 @@ export default async function NotificationPage() {
       <NotificationHeader />
       <QueryAsyncBoundary rejectedFallback={RejectedFallback} pendingFallback={<Loading />}>
         <PageAnimation>
-          {/* <HydrationProvider queryFn={getNotification} queryKey={Keys.notifications()}> */}
+          {/* <HydrationProvider queryFn={getNotification} queryKey={Keys.getNotifications()}> */}
           <NotificationSection />
           {/* </HydrationProvider> */}
         </PageAnimation>
